[REFACTOR] PHPStan: group legacy-UI custom rules under a LegacyUI namespace
Move the legacy-UI usage rules from scripts/PHPStan/Rules into a dedicated
Rules/LegacyUI subdirectory and namespace (ILIAS\Scripts\PHPStan\Rules\LegacyUI),
and update legacy_ui.neon to the new fully-qualified class names. Pure refactor,
no behavioural change.
