+++ scripts/PHPStan/Rules/LegacyUI/NoLegacyButtonUsagesRule.php
@@ -16,7 +16,7 @@
  *
  *********************************************************************/
 
-namespace ILIAS\Scripts\PHPStan\Rules;
+namespace ILIAS\Scripts\PHPStan\Rules\LegacyUI;
 
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\Rule;
@@ -24,21 +24,20 @@ use PhpParser\Node;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\RuleErrorBuilder;
 
-class NoLegacyModalUsagesRule extends LegacyClassUsageRule implements Rule
+class NoLegacyButtonUsagesRule extends LegacyClassUsageRule implements Rule
 {
     protected function getHumanReadableRuleName(): string
     {
-        return 'Legacy Modal Usages';
+        return 'Legacy Button Usages';
     }
 
     protected function getRelevantILIASVersion(): int
     {
-        return 10;
+        return 9;
     }
-
 
     protected function getForbiddenClasses(): array
     {
-        return ['ilModalGUI'];
+        return ['ilLinkButton'];
     }
 }
