+++ scripts/PHPStan/Rules/LegacyUI/NoLegacyExplorerUsagesRule.php
@@ -16,7 +16,7 @@
  *
  *********************************************************************/
 
-namespace ILIAS\Scripts\PHPStan\Rules;
+namespace ILIAS\Scripts\PHPStan\Rules\LegacyUI;
 
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\Rule;
@@ -24,11 +24,11 @@ use PhpParser\Node;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\RuleErrorBuilder;
 
-class NoLegacyToolbarUsagesRule extends LegacyClassUsageRule implements Rule
+class NoLegacyExplorerUsagesRule extends LegacyClassUsageRule implements Rule
 {
     protected function getHumanReadableRuleName(): string
     {
-        return 'Legacy Toolbar Usages';
+        return 'Legacy Explorer Usages';
     }
 
     protected function getRelevantILIASVersion(): int
@@ -36,13 +36,9 @@ class NoLegacyToolbarUsagesRule extends LegacyClassUsageRule implements Rule
         return 10;
     }
 
+
     protected function getForbiddenClasses(): array
     {
-        return ['ilToolbarGUI'];
-    }
-
-    protected function findMethodUsages(): bool
-    {
-        return true;
+        return ['ilExplorer', 'ilExplorerBaseGUI', 'ilTreeExplorerGUI'];
     }
 }
