+++ scripts/PHPStan/Rules/LegacyUI/NoLegacyPanelUsagesRule.php
@@ -16,7 +16,7 @@
  *
  *********************************************************************/
 
-namespace ILIAS\Scripts\PHPStan\Rules;
+namespace ILIAS\Scripts\PHPStan\Rules\LegacyUI;
 
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\Rule;
@@ -24,11 +24,11 @@ use PhpParser\Node;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\RuleErrorBuilder;
 
-class NoLegacyConfirmationUsagesRule extends LegacyClassUsageRule implements Rule
+class NoLegacyPanelUsagesRule extends LegacyClassUsageRule implements Rule
 {
     protected function getHumanReadableRuleName(): string
     {
-        return 'Legacy Confirmation Screen Usages';
+        return 'Legacy Panel Usages';
     }
 
     protected function getRelevantILIASVersion(): int
@@ -39,6 +39,6 @@ class NoLegacyConfirmationUsagesRule extends LegacyClassUsageRule implements Rul
 
     protected function getForbiddenClasses(): array
     {
-        return ['ilConfirmationGUI'];
+        return ['ilPanelGUI'];
     }
 }
