+++ scripts/PHPStan/Rules/LegacyUI/NoLegacyTableUsagesRule.php
@@ -16,7 +16,7 @@
  *
  *********************************************************************/
 
-namespace ILIAS\Scripts\PHPStan\Rules;
+namespace ILIAS\Scripts\PHPStan\Rules\LegacyUI;
 
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\Rule;
@@ -24,21 +24,21 @@ use PhpParser\Node;
 use PHPStan\Analyser\Scope;
 use PHPStan\Rules\RuleErrorBuilder;
 
-class NoLegacyCheckboxListUsagesRule extends LegacyClassUsageRule implements Rule
+class NoLegacyTableUsagesRule extends LegacyClassUsageRule implements Rule
 {
     protected function getHumanReadableRuleName(): string
     {
-        return 'Legacy Checkbox List Usages';
+        return 'Legacy Table Usages';
     }
 
     protected function getRelevantILIASVersion(): int
     {
-        return 9;
+        return 10;
     }
 
 
     protected function getForbiddenClasses(): array
     {
-        return ['ilCheckboxListOverlayGUI'];
+        return ['ilTable2GUI', 'ilTableGUI'];
     }
 }
